CategoryDetailCardView File
