--- TravelDiscovery/Discover/CategoryDetail/View/CategoryDetailCardImageView.swift
@@ -1,20 +1,0 @@
-//
-//  CategoryDetailCardImageView.swift
-//  TravelDiscovery
-//
-//  Created by Alejandro Robles on 23/06/23.
-//
-
-import SwiftUI
-
-struct CategoryDetailCardImageView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct CategoryDetailCardImageView_Previews: PreviewProvider {
-    static var previews: some View {
-        CategoryDetailCardImageView()
-    }
-}
